Create README.md
TASK-1) A product of our interest.and start exploring the product with our inference and summarize our findings. TASK-2) There are three application of complex nature using different literature. TASK-3) A case study of our own to solve a complex problem and apply the design thinking phases. TASK-4) AND fix a product configurator the knowledge.OR Develop a C-K theory for optimization.